feat: add quote visibility (#181)
* refactor(better-auth): narrow role type to 'user' | 'admin'

* fix: user role types and casl permissions

* feat: add migration for quote visibility field

* feat: add visibility field to quote types

* feat: implement `kyselyWhere`

* feat: filter quotes with `kyselyWhere`

* feat: add visibility select to Quote forms

* fix(authStore): remove unnecesary check if store is undefined

* chore(server): add `UserModule` with `getList` endpoint

* build: install combobox components

* chore(server): use userId to update & create quotes

* build: update shadcn

* refactor(shadcn): update ui components

* chore(client): implement `userApi`

* chore: update quotes using patch method

* chore: implement combobox select to choose user

* refactor: separate quoteForm to createQuoteForm and updateQuoteForm

* chore(server): validate quote update permissions before update

* feat: translate user label

* refactor: rename `useDebouncedValue` and `useComboboxAnchor` files

* refactor: remove unused `getUsersQuerySchema`

* refactor: make permissions a bit more clear that anyone can view public quotes

* refactor: optimize permission checks

* refactor: simplify sql query

* chore(selectUser): remove unused conditions

* chore: allow to getUserList to any user/guest

* fix(db): rename `post_visibility` to `quote_visibility`

* chore: remove 'use client' line

* chore: make quote `visibility` required

* refactor: visibility field translation

* Update apps/client/src/components/quote/form/update-quote-schema.ts


* Update apps/client/src/components/quote/form/create-quote-schema.ts


* chore(update-quote): make zod schema fields required

* refactor: separate quote-visibility schema

* Update apps/server/src/quote/quote.service.ts


* refactor: check quote update permission for every field

* chore: add permissions for `quote.service` methods

* refactor: use `createQuoteSchema`&`updateQuoteSchema` instead of `QuoteData`

* refactor: separate `quoteSchema` from `createQuoteSchema` and `updateQuoteSchema`

* refactor: translation structure

* chore: rename `create` endpoint to `createOwnQuote`

* chore: rename `UnauthorizedError` to `MissingUserError`

* chore(visibility-select): remove unused select placeholder

---------

+++ apps/server/src/quote/quote.service.ts
@@ -1,4 +1,6 @@
-import { ResultAsync } from 'neverthrow';
+import { errAsync, okAsync, ResultAsync } from 'neverthrow';
+import { AuthStore } from 'src/auth/auth-als.module';
+import { KyselyService } from 'src/database/kysely.service';
 import { QuoteId } from 'src/database/tables/quote.tables';
 import { QuoteListQueryDto } from 'src/quote/dto/quote-list-query.dto';
 import { UpdateQuoteDto } from 'src/quote/dto/update-quote.dto';
@@ -11,6 +13,7 @@ import {
   QuoteSearchQueryService,
   UpdateQuoteError,
 } from 'src/quote/quote.types';
+import { ForbiddenError, MissingUserError } from 'src/utils/errors/app-errors';
 
 import { Inject, Injectable } from '@nestjs/common';
 
@@ -25,10 +28,19 @@ export class QuoteService {
     private readonly quoteRepository: QuoteRepository,
     @Inject(QUOTE_SEARCH_QUERY_SERVICE)
     private readonly quoteSearchQueryService: QuoteSearchQueryService,
+    private readonly authStore: AuthStore,
+    private readonly db: KyselyService,
   ) {}
 
   getOne(id: QuoteId): ResultAsync<Quote, GetQuoteError> {
-    return this.quoteRepository.getOne(id);
+    const { ability } = this.authStore.getStore();
+    return this.quoteRepository.getOne(id).andThen((quote) => {
+      if (!ability.can('read', quote)) {
+        return errAsync(new ForbiddenError());
+      }
+
+      return okAsync(quote);
+    });
   }
 
   getList(
@@ -46,18 +58,51 @@ export class QuoteService {
     });
   }
 
-  create(data: CreateQuoteDto): ResultAsync<Quote, CreateQuoteError> {
-    return this.quoteRepository.create(data);
+  createOwnQuote(data: CreateQuoteDto): ResultAsync<Quote, CreateQuoteError> {
+    const { ability, user } = this.authStore.getStore();
+    if (!user) {
+      throw new MissingUserError();
+    }
+    if (!ability.can('create', 'Quote')) {
+      return errAsync(new ForbiddenError());
+    }
+
+    return this.quoteRepository.create(user.id, data);
   }
 
   update(
     id: QuoteId,
     data: UpdateQuoteDto,
   ): ResultAsync<Quote, UpdateQuoteError> {
-    return this.quoteRepository.update(id, data);
+    const { ability } = this.authStore.getStore();
+    const canUpdateEveryField = Object.keys(data).every((k) =>
+      ability.can('update', 'Quote', k),
+    );
+    if (!canUpdateEveryField) {
+      return errAsync(new ForbiddenError());
+    }
+
+    return this.db.withTransaction(() =>
+      this.quoteRepository.getOne(id).andThen((quote) => {
+        if (!ability.can('update', quote)) {
+          return errAsync(new ForbiddenError());
+        }
+
+        return this.quoteRepository.update(id, data);
+      }),
+    );
   }
 
   delete(id: QuoteId): ResultAsync<Quote, DeleteQuoteError> {
-    return this.quoteRepository.delete(id);
+    const { ability } = this.authStore.getStore();
+    return this.db.withTransaction(() =>
+      this.quoteRepository.getOne(id).andThen((quote) => {
+        if (!ability.can('delete', quote)) {
+          return errAsync(new ForbiddenError());
+        }
+
+        return this.quoteRepository.delete(id);
+      }),
+    );
   }
 }
